+++ wetterdienst/core/timeseries/summarize.py
@@ -12,8 +12,6 @@ from wetterdienst.metadata.resolution import Frequency
 from wetterdienst.util.logging import TqdmToLogger
 
 if TYPE_CHECKING:
-    from enum import Enum
-
     from wetterdienst.core.timeseries.request import TimeseriesRequest
     from wetterdienst.core.timeseries.result import StationsResult
 
@@ -66,13 +64,17 @@ def apply_station_values_per_parameter(
         ):
             log.info(f"Station for parameter {parameter.name} is too far away")
             continue
-        if parameter.name in param_dict and param_dict[parameter.name].finished:
+        if (parameter.dataset.name, parameter.name) in param_dict and param_dict[
+            (parameter.dataset.name, parameter.name)
+        ].finished:
             continue
         # Filter only for exact parameter
-        result_series_param = result_df.filter(pl.col(Columns.PARAMETER.value).eq(parameter.name))
+        result_series_param = result_df.filter(
+            pl.col(Columns.DATASET.value).eq(parameter.dataset.name), pl.col(Columns.PARAMETER.value).eq(parameter.name)
+        )
         if result_series_param.drop_nulls("value").is_empty():
             continue
-        if parameter.name not in param_dict:
+        if (parameter.dataset.name, parameter.name) not in param_dict:
             frequency = Frequency[parameter.dataset.resolution.value.name].value
             df = pl.DataFrame(
                 {
@@ -86,12 +88,14 @@ def apply_station_values_per_parameter(
                 },
                 orient="col",
             )
-            param_dict[parameter.name] = _ParameterData(df)
+            param_dict[(parameter.dataset.name, parameter.name)] = _ParameterData(df)
         result_series_param = (
-            param_dict[parameter.name].values.select("date").join(result_series_param, on="date", how="left")
+            param_dict[(parameter.dataset.name, parameter.name)]
+            .values.select("date")
+            .join(result_series_param, on="date", how="left")
         )
         result_series_param = result_series_param.get_column(Columns.VALUE.value).rename(station["station_id"])
-        extract_station_values(param_dict[parameter.name], result_series_param, True)
+        extract_station_values(param_dict[(parameter.dataset.name, parameter.name)], result_series_param, True)
 
 
 def calculate_summary(stations_dict: dict, param_dict: dict) -> pl.DataFrame:
@@ -99,6 +103,7 @@ def calculate_summary(stations_dict: dict, param_dict: dict) -> pl.DataFrame:
         pl.DataFrame(
             schema={
                 Columns.DATE.value: pl.Datetime(time_zone="UTC"),
+                Columns.DATASET.value: pl.String,
                 Columns.PARAMETER.value: pl.String,
                 Columns.VALUE.value: pl.Float64,
                 Columns.DISTANCE.value: pl.Float64,
@@ -106,14 +111,15 @@ def calculate_summary(stations_dict: dict, param_dict: dict) -> pl.DataFrame:
             },
         ),
     ]
-    for parameter, param_data in param_dict.items():
+    for (dataset, parameter), param_data in param_dict.items():
         param_df = pl.DataFrame({Columns.DATE.value: param_data.values.get_column(Columns.DATE.value)})
         results = []
         for row in param_data.values.select(pl.all().exclude("date")).iter_rows(named=True):
-            results.append(apply_summary(row, stations_dict, parameter))
+            results.append(apply_summary(row, stations_dict, dataset, parameter))
         results = pl.DataFrame(
             results,
             schema={
+                Columns.DATASET.value: pl.String,
                 Columns.PARAMETER.value: pl.String,
                 Columns.VALUE.value: pl.Float64,
                 Columns.DISTANCE.value: pl.Float64,
@@ -127,6 +133,7 @@ def calculate_summary(stations_dict: dict, param_dict: dict) -> pl.DataFrame:
     df = df.with_columns(pl.col(Columns.VALUE.value).round(2), pl.col(Columns.DISTANCE.value).round(2))
     return df.sort(
         by=[
+            Columns.DATASET.value,
             Columns.PARAMETER.value,
             Columns.DATE.value,
         ],
@@ -136,12 +143,13 @@ def calculate_summary(stations_dict: dict, param_dict: dict) -> pl.DataFrame:
 def apply_summary(
     row: dict,
     stations_dict: dict,
-    parameter: Enum,
-) -> tuple[Enum, float | None, float | None, str | None]:
+    dataset: str,
+    parameter: str,
+) -> tuple[str, str, float | None, float | None, str | None]:
     vals = {s: v for s, v in row.items() if v is not None}
     if not vals:
-        return parameter, None, None, None
+        return dataset, parameter, None, None, None
     value = list(vals.values())[0]
     station_id = list(vals.keys())[0][1:]
     distance = stations_dict[station_id][2]
-    return parameter, value, distance, station_id
+    return dataset, parameter, value, distance, station_id