+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Stripes: Replace matplotlib by plotly
 - Explorer: Add download button for plot
 - Split up plotting extras into `plotting` and `matplotlib`
+- Interpolation/Summary: Add dataset to DataFrame
 
 ### Refactor
 - Update docker image extras